Skip cover_db/, vim backup and .swp files from MANIFEST
Also skip the MANIFEST.SKIP file itself: the `03-manifest.t` test which
checks the manifest was complaining that the MANIFEST.SKIP wasn't in the
MANIFEST, which it doesn't need to be (and was in fact removed from the
MANIFEST in a recent commit), hence the MANIFEST.SKIP should also skip
itself.  Now the RELEASE_TESTING tests pass.