Use constants now exported from Mojo::WebSocket
The constants `DEBUG`, `TEXT`, and `CLOSE` were (in Mojolicious versions
below 6.40) defined in `Mojo::Transaction::WebSocket`.  They are now (in
versions 6.40 and above) defined in `Mojo::WebSocket`, also the `TEXT` and
`CLOSE` constants have been renamed with a `WS_` prefix.  The `BINARY`
constnat has been removed since it turns out this wasn't used in
`MojoX::Transaction::WebSocket76`.  This commit updates the code to use the
new constant names and thus allows the module to work with Mojolicious
versions 6.40 and above.